requirements.txt
python-telegram-bot==20.7
pandas==2.2.2
yfinance==0.2.40
numpy==1.26.4
pytz==2024.1
